rework: Organize dashboard components, improve delete UI
