+++ client/src/components/dashboard/quiz/buttons/DeleteQuizButton.jsx
@@ -1,25 +1,29 @@
-import { toast } from "sonner";
-import { deleteQuizById, getAllQuizzes } from "../../api/auth";
+import { truncateText } from "../../../../utility/truncate";
+import { deleteQuizById, getAllQuizzes } from "../../../../api/auth";
 
-export function DeleteQuizButton({ id, quizName, description, setQuizzes }) {
-  function handleClick(e) {
+import { toast } from "sonner";
+
+export function DeleteQuizButton({ id, quizName, setQuizzes }) {
+  const handleClick = (e) => {
     e.stopPropagation();
     document.getElementById("delete_quiz_modal").showModal();
   }
 
-  function handleClose() {
+  const handleClose = () => {
     document.getElementById("delete_quiz_modal").close();
   }
 
-  async function handleSubmit(e) {
+  const handleSubmit = async (e) => {
     e.preventDefault();
 
     try {
       const data = await deleteQuizById(id);
-      toast.success(data.message);
-      handleClose();
       const quizzes = await getAllQuizzes();
+
       setQuizzes(quizzes.data);
+      handleClose();
+
+      toast.success(data.message);
     } catch (error) {
       toast.error(error.message);
     }
@@ -32,13 +36,14 @@ export function DeleteQuizButton({ id, quizName, description, setQuizzes }) {
       </button>
       <dialog
         id="delete_quiz_modal"
-        onClick={(e) => e.stopPropagation()}
         className="modal"
+        onClick={(e) => e.stopPropagation()}
       >
         <div className="modal-box">
-          <h3 className="font-bold text-lg">Are you sure you want to delete</h3>
-          <p className="py-4">{quizName}</p>
-          <p>{description}</p>
+            <h3 className="font-bold text-lg" title={quizName}>
+              Are you sure you want to delete {truncateText(quizName, 15)}
+            </h3>
+          <p className="text-center pt-4 font-bold">This action is irreversible!</p>
           <div className="modal-action">
             <form onSubmit={handleSubmit}>
               <div className="modal-action">
@@ -49,8 +54,10 @@ export function DeleteQuizButton({ id, quizName, description, setQuizzes }) {
                 >
                   Cancel
                 </button>
+
+                {/* TODO: Maybe an icon here */}
                 <button type="submit" className="btn btn-error">
-                  Delete Quiz
+                  Yes I know what I'm doing
                 </button>
               </div>
             </form>
